[CLI] Replace MessageCollector usage in cli-native(-klib)
#KT-78277
